Sharding Fixes #20
