+++ package.json
@@ -14,6 +14,7 @@
     "@types/signale": "^1.4.2",
     "chalk": "4.1.2",
     "discord-api-types": "^0.25.2",
+    "discord-hybrid-sharding": "^1.3.6",
     "discord.js": "^13.3.1",
     "dotenv": "^10.0.0",
     "lodash": "^4.17.21",
